test(chat/infra): extract HubConnectionHelper from ChatHubTests
